settings: Explain the build status on our dialog
This is for those who asks "WUT IS SUSHIFIRED SUR" or stuff like that. So, let's explain it.
On our previous iteration we did this string hardcoded so, now let's make it translatable, really unneeded but eh

Thanks google for the regrettable_lack_of_easter_egg string and AndroidPolice to point it out.
	Source for the shrug: https://android.googlesource.com/platform/frameworks/base/+/android-8.0.0_r1/packages/SystemUI/res/values/strings.xml#1322